Update __init__.py
Updated add keymap to install in to the WINDOW entry. The key to use is  J (For Jog, like MotionBuilder, though the Maya hotkey is K) . Note that if you are in Grease Pencil DRAW mode, you need to hit this J key to activate scrub in a timeline or other window as Grease Pencil DRAW overrides the hotkey. You can assign it to another KEY or add a modifier to it as a workaround.